Added shrinkpdf.sh script to reduce pdf size.
I also notice that there is a previous script to do the same,
probably done by me by extracting the useful part from this new
script, that is new here but old in the web (so I surely came
accross it before).
Also added a README in the bin_files folder to keep a neat list
with comments.